moving code around, linting
